+++ src/main/java/com/samyookgoo/palgoosam/notification/controller/NotificationController.java
@@ -1,7 +1,8 @@
 package com.samyookgoo.palgoosam.notification.controller;
 
+import com.samyookgoo.palgoosam.auction.repository.AuctionRepository;
 import com.samyookgoo.palgoosam.common.response.BaseResponse;
-import com.samyookgoo.palgoosam.notification.dto.FcmTokenSaveRequestDto;
+import com.samyookgoo.palgoosam.notification.fcm.dto.FcmTokenSaveRequestDto;
 import com.samyookgoo.palgoosam.notification.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class NotificationController {
 
     private final NotificationService notificationService;
+    private final AuctionRepository auctionRepository;
 
     @PostMapping("/fcm-token")
     public ResponseEntity<BaseResponse> saveFcmToken(@RequestBody FcmTokenSaveRequestDto requestDto) {
