+++ src/main/java/com/samyookgoo/palgoosam/notification/fcm/dto/FcmMessageData.java
@@ -1,4 +1,4 @@
-package com.samyookgoo.palgoosam.notification.dto;
+package com.samyookgoo.palgoosam.notification.fcm.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
