+++ src/main/java/com/samyookgoo/palgoosam/notification/fcm/service/FcmInitialization.java
@@ -1,4 +1,4 @@
-package com.samyookgoo.palgoosam.notification.service;
+package com.samyookgoo.palgoosam.notification.fcm.service;
 
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
