+++ src/main/java/com/samyookgoo/palgoosam/notification/service/NotificationService.java
@@ -6,14 +6,16 @@ import com.samyookgoo.palgoosam.auction.repository.AuctionImageRepository;
 import com.samyookgoo.palgoosam.auction.repository.AuctionRepository;
 import com.samyookgoo.palgoosam.common.response.BaseResponse;
 import com.samyookgoo.palgoosam.notification.constant.NotificationStatusType;
-import com.samyookgoo.palgoosam.notification.constant.SubscriptionType;
 import com.samyookgoo.palgoosam.notification.domain.NotificationHistory;
 import com.samyookgoo.palgoosam.notification.domain.NotificationStatus;
-import com.samyookgoo.palgoosam.notification.dto.FcmMessageData;
-import com.samyookgoo.palgoosam.notification.dto.FcmNotificationRequestDto;
-import com.samyookgoo.palgoosam.notification.dto.FcmTokenSaveRequestDto;
+import com.samyookgoo.palgoosam.notification.fcm.dto.FcmMessageData;
+import com.samyookgoo.palgoosam.notification.fcm.dto.FcmNotificationRequestDto;
+import com.samyookgoo.palgoosam.notification.fcm.dto.FcmTokenSaveRequestDto;
+import com.samyookgoo.palgoosam.notification.fcm.service.FirebaseCloudMessageService;
 import com.samyookgoo.palgoosam.notification.repository.NotificationHistoryRepository;
 import com.samyookgoo.palgoosam.notification.repository.NotificationStatusRepository;
+import com.samyookgoo.palgoosam.notification.subscription.constant.SubscriptionType;
+import com.samyookgoo.palgoosam.notification.subscription.service.AuctionSubscriptionService;
 import com.samyookgoo.palgoosam.user.domain.User;
 import com.samyookgoo.palgoosam.user.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
