+++ src/main/java/com/samyookgoo/palgoosam/notification/subscription/constant/SubscriptionType.java
@@ -1,4 +1,4 @@
-package com.samyookgoo.palgoosam.notification.constant;
+package com.samyookgoo.palgoosam.notification.subscription.constant;
 
 public enum SubscriptionType {
     BIDDER,
