+++ src/main/java/com/samyookgoo/palgoosam/notification/subscription/repository/AuctionSubscriptionRepository.java
@@ -1,7 +1,7 @@
-package com.samyookgoo.palgoosam.notification.repository;
+package com.samyookgoo.palgoosam.notification.subscription.repository;
 
 import com.samyookgoo.palgoosam.auction.domain.Auction;
-import com.samyookgoo.palgoosam.notification.domain.AuctionSubscription;
+import com.samyookgoo.palgoosam.notification.subscription.domain.AuctionSubscription;
 import com.samyookgoo.palgoosam.user.domain.User;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
